Fix behavior around static-but-growable strings
If a string is declared as VS_TYPE_STATIC but also has VS_TYPE_GROWABLE,
it should be possible to upgrade that string to a growable string. Once
the string is upgraded, it must be freed with `vs_deinit` as a dynamic
string would be. Add tests to this effect as well.

Also remove a rather unprofessional debug printf that was somehow added
in a previous commit.